Extract code from `ActionElement` to `ExtraActions` to simplify file.
